+++ app/src/store/slices/tabs-layout/initialState.ts
@@ -1,0 +1,8 @@
+import { TabsLayoutState } from "./types";
+
+export const initialState: TabsLayoutState = {
+  apiClient: {
+    tabs: [],
+    activeTab: null,
+  },
+};
