+++ app/src/store/slices/tabs-layout/slice.ts
@@ -1,7 +1,8 @@
 import { createSlice } from "@reduxjs/toolkit";
 import { ReducerKeys } from "store/constants";
 import getReducerWithLocalStorageSync from "store/getReducerWithLocalStorageSync";
-import { addTab, initialState, resetState, setActiveTab, setTabs, updateTab } from "./reducers";
+import { addTab, resetState, setActiveTab, setTabs, updateTab } from "./case-reducers";
+import { initialState } from "./initialState";
 
 const slice = createSlice({
   name: ReducerKeys.TABS_LAYOUT,
@@ -15,10 +16,8 @@ const slice = createSlice({
   },
 });
 
-export const { actions, reducer } = slice;
-
+const { actions, reducer } = slice;
 export const tabsLayoutActions = actions;
-
 export const tabsLayoutReducerWithLocalSync = getReducerWithLocalStorageSync(ReducerKeys.TABS_LAYOUT, reducer, [
   "apiClient",
 ]);
