+++ app/src/store/slices/tabs-layout/types.ts
@@ -1,0 +1,9 @@
+import { TabsLayout } from "layouts/TabsLayout";
+
+export type TabsLayoutState = Record<
+  string,
+  {
+    tabs: TabsLayout.Tab[];
+    activeTab: TabsLayout.Tab | null;
+  }
+>;
